Time: 0 ms (100%), Space: 8.3 MB (5.09%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -5,14 +5,14 @@ public:
         int end = x,mid,ans = 0;
         while(start <= end){
             mid = start + (end - start)/2;
-            if(((long)mid*mid) > x){
+            if(((long)mid*mid) > (long)x){
                 end = mid - 1;
             }
-            else if(((long)mid*mid) < x){
+            else if(((long)mid*mid) < (long)x){
                 ans = mid;
                 start = mid + 1;
             }
-            if((mid*mid) == x){
+            if(((long)mid*mid) == x){
                 ans = mid;
                 break;
             }
